Use maximumBy
Yes, maximumBy is a partial function.
My reasoning is that it makes more sense
to throw an exception than to return "Criminal not found".
A caller might use the return value to select an account to operate on,
so if we return "Criminal not found",
that would be be undesirable.